pyproject.toml
[tool.poetry]
name = "gestao_clientes"
version = "0.1.0"
description = "Client management project created in Django Framework."
authors = ["Wesley Mendes <wesley@wesleymendes.com.br>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "2.2.6"
boto3 = "1.5.36"
botocore = "1.8.50"
dj-database-url = "0.4.2"
dj-static = "0.0.6"
django-bootstrap-form = "3.3"
django-storages = "1.6.5"
docutils = "0.14"
gunicorn = "^20.0.4"
jmespath = "0.9.3"
Pillow = "5.0.0"
python-dateutil = "2.6.1"
python-decouple = "3.1"
pytz = "2020.5"
s3transfer = "0.1.13"
six = "1.11.0"
static3 = "0.7.0"
psycopg2-binary = "^2.8.6"
django-extensions = "^3.1.0"
django-debug-toolbar = "^3.2"

[tool.poetry.dev-dependencies]
mypy = "^0.790"
black = "^20.8b1"
factory-boy = "^3.1.0"
Faker = "^5.0.2"
flake8 = "^3.8.4"
ipdb = "^0.13.4"
pytest = "^6.2.1"
pytest-django = "^4.1.0"
pytest-flake8 = "^1.0.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
